pyproject.toml
[project]
name = "zettelvault"
version = "0.1.0"
description = "Transform an Obsidian vault into PARA + Zettelkasten structure using DSPy and RLM"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dspy>=3.1.3",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0",
]

[project.scripts]
zettelvault = "zettelvault.__main__:main"

[dependency-groups]
dev = [
    "pip-audit>=2.10.0",
    "ruff>=0.15.6",
]
